+++ internal/storage/postgres/migrations/00001_init.up.sql
@@ -4,8 +4,8 @@ CREATE TABLE users (
     id UUID PRIMARY KEY DEFAULT gen_random_uuid(),
     login VARCHAR(200) UNIQUE NOT NULL,
     hash_password VARCHAR(200) NOT NULL,
-    balance int,
-    withdrawn int
+    balance REAL,
+    withdrawn REAL
 );
 
 CREATE TYPE status AS ENUM ('NEW', 'PROCESSING', 'INVALID', 'PROCESSED');
@@ -14,7 +14,7 @@ CREATE TABLE orders (
     id bigint PRIMARY KEY UNIQUE,
     status status NOT NULL ,
     created_at TIMESTAMP  DEFAULT now(),
-    accrual INT,
+    accrual REAL,
     username VARCHAR(200),
     CONSTRAINT counter_positive_check CHECK (accrual::numeric >= 0),
     FOREIGN KEY(username) references users(login)
